backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
flask-cors==4.0.0
python-dotenv==1.0.0
werkzeug==2.3.7
pytest==7.4.3
black==23.11.0
flake8==6.1.0